requirements.txt
Flask==3.0.0
pdfplumber==0.11.0
Werkzeug==3.0.1
python-dotenv==1.2.1
gunicorn==21.2.0
stripe==8.0.0
requests==2.31.0
urllib3==1.26.18
openai>=2.0.0
weasyprint==59.0
pydyf==0.10.0
